ThemePref – write offset data to disk
Requested by Pluto.  This adds a ThemePref for "Write Offset Data to Disk."

If enabled, it writes a txt file to ./Simply Love/OffsetData/ after each song with per-column offset data.  The hope is that this helps diagnose and fix input/timing issues.

This has only been tested in dance gameplay, and only supports English for now.